Ensure that class extends and throws type annotations use u2, not u1.

+++ src/share/classes/com/sun/tools/classfile/ClassWriter.java
@@ -770,11 +770,11 @@ public class ClassWriter {
                 break;
             // class extends or implements clause
             case CLASS_EXTENDS:
-                out.writeByte(p.type_index);
+                out.writeShort(p.type_index);
                 break;
             // throws
             case THROWS:
-                out.writeByte(p.type_index);
+                out.writeShort(p.type_index);
                 break;
             // method parameter
             case METHOD_FORMAL_PARAMETER:
